Simplification: Bug Zero Charge on Check Out
This test is carried out to replicate the bug that cause the service charge to be zero while performing the hotel check out use case. The test is performed without any user input and all the methods call are automated. As there is no method that can be used to check the charge of the room without changing the code, the print output is taken as a confirmation of the bugs that has been reported.