Simplification: Bug Zero Charge on Check Out
This document presents a screenshot of simplified code that can generate the bugs that are presented for debugging.  The test is performed without any user input and all the methods call are automated. As there is no method that can be used to check the charge of the room without changing the code, the print output is taken as a confirmation of the bugs that has been reported.