Simplification: Bug Charge Checked out Room
As, the purpose of this test is to simplify the reproduction of the bug using an automated test without any user input.  The test is performed without any user input and all the methods call are automated. As there is no method that can be used to check the charge of the room without changing the code, the print output is taken as a confirmation of the bugs that has been reported.