Tracing: Bug Charge Checked out Room
As bug that relates to being able to add service charge to room that has already checked out, and by looking at the RecordServiceCTL , there appear to be problem on method that is supposed to remove the booking. And the error was trace as it progressed which can be seen in Hypothesis that is being tested.